Module rustymq::base::message

Message module

Datastructures, trait definitions for Message handling within RustyMQ

Details

This module containts the definition of Messages, which are a core concept of RustyMQ. All the datastructures, traits and enums needed to compose messages for sending and then receiving are defined here

Structs

MessageMetadata
Metadata for every ZeroMQ Message

RawMessage
RawMessage is a message type implementing the Message trait, with a binary buffer set as payload

TypedMessage
Typed message is a generic implementation for a Message trait with custom type

Enums

Part
Multipart message part identifier

PartError
Multipart message part tracking error

Traits

Message
RustyMQ Message Trait for a universal interface for messages with a given associated payload type

SerializableMessagePayload
Implementing this trait on a struct allows for using a TypedMessage with the struct

Type Definitions

Buffer
General buffer type to be used in RusyMQ

BufferMutSlice
General buffer mutable slice type to be used in RustyMQ

BufferSlice
General buffer slice type to be used in RustyMQ

CommunicationModelId
Type definition for the identifier for the different communication models. For details see super::super::model:here

ConversationId
Type definition for conversation identifier data structure

MessageId
Type definition for message identifier data structure

PeerId
Type definition for conversation peer identifier data structure