Enum rustymq::base::message::Part

pub enum Part {
    Single,
    Intermediate(u32),
    Final(u32),
}

Multipart message part identifier

Enum for tracking multipart messages

Description

Enum to track multiplart message sequences more easily. Although the iteration of these should be handled manually by the user

Variants

Single

Signaling this is a single message

Intermediate(u32)

Signaling this is an intermediate message of a multipart message seqence and the index

Final(u32)

Signaling this is a final message of a multipart message sequence and the index

Implementations

impl Part

pub fn single() -> Self

Generate new instance for a single message without multiple aprts

Example

let original = Part::single();
assert!(std::matches!(original, Part::Single));

pub fn start_multipart() -> Self

Generate a new instance for trackinga multipart message

Example

let original = Part::start_multipart();
assert!(std::matches!(original, Part::Intermediate(0)));

pub fn is_multipart(&self) -> bool

Query if part is multipart

pub fn is_continueable(&self) -> bool

Query if continuable part

pub fn is_initial(&self) -> bool

Query if part can be used as a first message

pub fn is_last(&self) -> bool

Query if part can be used as the last message

pub fn next_multipart(self) -> Result<Self, PartError>

Generate new instance for multipart message metadata with advancing the part id by one

Example

let original = Part::start_multipart();
assert!(std::matches!(original, Part::Intermediate(0)));

let result = original.next_multipart().unwrap();
assert!(std::matches!(result, Part::Intermediate(1)));

pub fn as_final_multipart(self) -> Result<Self, PartError>

Generate new instance keeping the part index but change to final message

Example

let original = Part::start_multipart();
assert!(std::matches!(original, Part::Intermediate(0)));

let result = original.as_final_multipart().unwrap();
assert!(std::matches!(result, Part::Final(0)));

pub fn next_final_multipart(self) -> Result<Self, PartError>

Generate new instance keeping the part index but change to final message

Example

let original = Part::start_multipart();
assert!(std::matches!(original, Part::Intermediate(0)));

let result = original.next_final_multipart().unwrap();
assert!(std::matches!(result, Part::Final(1)));

pub fn update_to_next_part(&mut self, other: &Self) -> Result<(), PartError>

Update to next index if possible. This is used to update the part index to the one in other if it is the next consecutive one

Examples

let mut original = Part::start_multipart();

assert!(std::matches!(original, Part::Intermediate(0)));
original.update_to_next_part(&Part::Intermediate(1)).unwrap();

assert!(std::matches!(original, Part::Intermediate(1)));

let mut original = Part::start_multipart();

assert!(std::matches!(original, Part::Intermediate(0)));
assert!(std::matches!(original.update_to_next_part(&Part::Intermediate(2)),
                      Err(PartError::NotConsecutivePart)));

let mut original = Part::Final(2);

assert!(std::matches!(original.update_to_next_part(&Part::Intermediate(0)),
                      Err(PartError::AlreadyFinishedMultipart)));

pub fn to_next_part(self, other: &Self) -> Result<Self, PartError>

Generate new instance with next index if possible from other. Same as update_to_next_part function without mutability