Trait rustymq::base::message::Message

pub trait Message: Sized {
    type Payload;
    fn new(payload: Self::Payload) -> Self;

    fn with_metadata(metadata: MessageMetadata, payload: Self::Payload) -> Self;

    fn metadata(&self) -> &MessageMetadata;

    fn into_metadata(self) -> MessageMetadata;

    fn payload(&self) -> &Self::Payload;

    fn into_payload(self) -> Self::Payload;

    fn into_parts(self) -> (MessageMetadata, Self::Payload);

    fn mutated_metadata<Mutator: Fn(MessageMetadata) -> MessageMetadata>(
        self, 
        mutator: Mutator
    ) -> Self { ... }

    fn message_id(&self) -> &Option<MessageId> { ... }

    fn conversation_id(&self) -> &Option<ConversationId> { ... }

    fn communication_model_id(&self) -> &Option<CommunicationModelId> { ... }

    fn peer_id(&self) -> &Option<PeerId> { ... }

    fn is_multipart(&self) -> bool { ... }

    fn part(&self) -> &Part { ... }

    fn apply_message_id(self, message_id: MessageId) -> Self { ... }

    fn apply_random_message_id(self) -> Self { ... }

    fn ensure_random_message_id(self) -> Self { ... }

    fn apply_conversation_id(self, conversation_id: ConversationId) -> Self { ... }

    fn apply_random_conversation_id(self) -> Self { ... }

    fn ensure_random_conversation_id(self) -> Self { ... }

    fn apply_peer_id(self, peer_id: PeerId) -> Self { ... }

    fn apply_communication_model_id(
        self, 
        model_id: CommunicationModelId
    ) -> Self { ... }

    fn continue_conversation(self, payload: Self::Payload) -> Self
    where
        Self: Sized,
    { ... }

    fn continue_conversation_from_metadata(
        self, 
        metadata: MessageMetadata
    ) -> Self
    where
        Self: Sized,
    { ... }
}

RustyMQ Message Trait for a universal interface for messages with a given associated payload type

Associated Types

type Payload

Associated type of the payload type

Loading content...
Required methods

fn new(payload: Self::Payload) -> Self

Create a new message with a the given payload

fn with_metadata(metadata: MessageMetadata, payload: Self::Payload) -> Self

Create a new message with the given metadata and payload

fn metadata(&self) -> &MessageMetadata

Query the metadata stored in the message

fn into_metadata(self) -> MessageMetadata

Consume message and return metadata

fn payload(&self) -> &Self::Payload

Query the payload of the message

fn into_payload(self) -> Self::Payload

Consume message and return payload

fn into_parts(self) -> (MessageMetadata, Self::Payload)

Consume message and return metadata and payload

Loading content...
Provided methods

fn mutated_metadata<Mutator: Fn(MessageMetadata) -> MessageMetadata>(
self,
mutator: Mutator
) -> Self

Generate new instance with mutated metadata but keeping the payload.

fn message_id(&self) -> &Option<MessageId>

Query the message identifier from the stored metadata

fn conversation_id(&self) -> &Option<ConversationId>

Query the conversation identifier from the stored metadata

fn communication_model_id(&self) -> &Option<CommunicationModelId>

Query the communication model identifier from the stored metadata if present

fn peer_id(&self) -> &Option<PeerId>

Query the peer identifier from the stored metadata if present

fn is_multipart(&self) -> bool

Query if the stored metadata is multipart

fn part(&self) -> &Part

Query if part tracker from the stored metadata

fn apply_message_id(self, message_id: MessageId) -> Self

Generate new instance with new message identifier in the stored metadata but keeping the payload

fn apply_random_message_id(self) -> Self

Generate new instance with new random message identifier in the stored metadata but keeping the payload

fn ensure_random_message_id(self) -> Self

Generate new instance with with ensured random message identifier (kept if already set) in the stored metadata but keeping the payload

fn apply_conversation_id(self, conversation_id: ConversationId) -> Self

Generate new instance with new conversation identifier in the stored metadata but keeping the payload

fn apply_random_conversation_id(self) -> Self

Generate new instance with new random conversation identifier in the stored metadata but keeping the payload

fn ensure_random_conversation_id(self) -> Self

Generate new instance with with ensured random conversation identifier (kept if already set) in the stored metadata but keeping the payload

fn apply_peer_id(self, peer_id: PeerId) -> Self

Generate new instance with new peer identifier in the stored metadata but keeping the payload

fn apply_communication_model_id(self, model_id: CommunicationModelId) -> Self

Generate new instance with new communication model identifier in metadata but keeping everything else

fn continue_conversation(self, payload: Self::Payload) -> Self where
Self: Sized,

Generate new instance by continuing metadata exchange and with the given payload

fn continue_conversation_from_metadata(self, metadata: MessageMetadata) -> Self where
Self: Sized,

Generate new instance by keeping payload and continuing metadata exchange

Loading content...