Enum rustymq::base::socket::SocketError

pub enum SocketError {
    IncorrectStateError,
    UnrelatedPeer,
    UnrelatedConversation,
    UnrelatedConversationPart,
    QueueDepthReached,
    TransportIntegrityError,
    UnknownPeer,
    TransportMethodAlreadyInUse,
    TransportTargetUnreachable,
    Disconnected,
    ConnectionRefused,
    Timeout,
    IncompatiblePeer,
    HandshakeFailed,
    UnsupportedOpFlag(OpFlag),
    MissingDNSDomain,
    InvalidTransportMethod,
    AlreadyConnected,
    NotSupportedOperation,
    CouldNotConnect,
}

Socket errors

Contains the possible errors for a socket that may need to be handled on the user side

Variants

IncorrectStateError

Signals if the socket is not in the correct state to process the requested operation

UnrelatedPeer

The specified peer identifier in MessageMetadata is unkown by the

UnrelatedConversation

The specified conversation in the MessageMetadata is not known by the socket

UnrelatedConversationPart

The specified conversation part in the MessageMetadata is not known by the socket

QueueDepthReached

The internal gueue in Socket or Transport has reached its limit

TransportIntegrityError

Integrity checks on transport layer failed

UnknownPeer

The target peer identifier is not specified in the MessageMetadata and cannot be inferred

TransportMethodAlreadyInUse

The specific TransportMethod is already used by the socket

TransportTargetUnreachable

The specific TransportMethod cannot be found, cannot be connected to

Disconnected

The tharget peer specified in MessageMetadata disconnected

ConnectionRefused

The tharget specified in TransportMethod refused the connection

Timeout

The operation timed out

IncompatiblePeer

The connected peer has an incompatible version or communication model

HandshakeFailed

Handshake with the target peer failed

UnsupportedOpFlag(OpFlag)

The OpFlag specified for the InwardSocket or OutwardSocket operation is not supprted by the socket or the underlieing Transport

MissingDNSDomain

DNS domain name required but not available in NetworkAddress

InvalidTransportMethod

The TransportMethod is not supported by the Transport

AlreadyConnected

The connection already exists

NotSupportedOperation

The requested operation is not supported by socket

CouldNotConnect

Could not connect to address provided by TransportMethod