Trait rustymq::base::socket::BidirectionalSocket

pub trait BidirectionalSocket: OutwardSocket + InwardSocket {
    fn query(
        &self, 
        message: RawMessage, 
        send_flags: OpFlag, 
        recieve_flags: OpFlag
    ) -> Result<RawMessage, (Option<PeerId>, SocketError)> { ... }

    fn query_typed<S, R>(
        &self, 
        message: TypedMessage<S>, 
        send_flags: OpFlag, 
        recieve_flags: OpFlag
    ) -> Result<TypedMessage<R>, QueryTypedError<S, R>>
    where
        S: SerializableMessagePayload,
        R: SerializableMessagePayload,
    { ... }

    fn respond<T: Fn(RawMessage) -> RawMessage>(
        &self, 
        receive_flags: OpFlag, 
        send_flags: OpFlag, 
        processor: T
    ) -> Result<(), (Option<PeerId>, SocketError)> { ... }

    fn respond_typed<R, S, Q: Fn(TypedMessage<R>) -> TypedMessage<S>>(
        &self, 
        receive_flags: OpFlag, 
        send_flags: OpFlag, 
        processor: Q
    ) -> Result<(), QueryTypedError<S, R>>
    where
        R: SerializableMessagePayload,
        S: SerializableMessagePayload,
    { ... }
}

Bidirectional Socket trait

The main interface for the user to send and receive messages through the same opened RustyMQ socket There can be several implementations of it depending on the communication model used

Provided methods

fn query(
&self,
message: RawMessage,
send_flags: OpFlag,
recieve_flags: OpFlag
) -> Result<RawMessage, (Option<PeerId>, SocketError)>

Execute a query (send and then receive) with a raw message This does not guaratee that the response is for the same message that was sent

fn query_typed<S, R>(
&self,
message: TypedMessage<S>,
send_flags: OpFlag,
recieve_flags: OpFlag
) -> Result<TypedMessage<R>, QueryTypedError<S, R>> where
S: SerializableMessagePayload,
R: SerializableMessagePayload,

Execute a query (send and then receive) with a typed message. Requires S and R to be convertibe to implement SerializableMessagePayload

fn respond<T: Fn(RawMessage) -> RawMessage>(
&self,
receive_flags: OpFlag,
send_flags: OpFlag,
processor: T
) -> Result<(), (Option<PeerId>, SocketError)>

Respont to a query (receive and then reply) with a raw message

fn respond_typed<R, S, Q: Fn(TypedMessage<R>) -> TypedMessage<S>>(
&self,
receive_flags: OpFlag,
send_flags: OpFlag,
processor: Q
) -> Result<(), QueryTypedError<S, R>> where
R: SerializableMessagePayload,
S: SerializableMessagePayload,

Respond to a query (receive and then reply) with a typed message. Requires S and R to be convertibe to implement SerializableMessagePayload

Loading content...