Trait rustymq::base::socket::OutwardSocket

pub trait OutwardSocket: Socket {
    fn send(
        &self, 
        message: RawMessage, 
        flags: OpFlag
    ) -> Result<MessageMetadata, SocketError>;

    fn send_typed<T>(
        &self, 
        message: TypedMessage<T>, 
        flags: OpFlag
    ) -> Result<MessageMetadata, SendTypedError<T>>
    where
        T: SerializableMessagePayload,
    { ... }
}

Outward Socket trait

The main interface for the user to send messages through the opened RustyMQ socket There can be several implementations of it depending on the communication model used

Required methods

fn send(
&self,
message: RawMessage,
flags: OpFlag
) -> Result<MessageMetadata, SocketError>

Send a raw message

Loading content...
Provided methods

fn send_typed<T>(
&self,
message: TypedMessage<T>,
flags: OpFlag
) -> Result<MessageMetadata, SendTypedError<T>> where
T: SerializableMessagePayload,

Send a typed message. Requires T to implement the SerializableMessagePayload trait

Loading content...