Trait rustymq::base::socket::Socket

pub trait Socket: Send + Sync {
    fn connect(
        &self, 
        target: TransportMethod
    ) -> Result<Option<PeerId>, SocketError>;

    fn bind(
        &self, 
        target: TransportMethod
    ) -> Result<Option<PeerId>, SocketError>;

    fn close_connection(
        &self, 
        peer_identification: PeerIdentification
    ) -> Result<(), SocketError>;

    fn close(self) -> Result<(), SocketError>;
}

General Socket trait

The main interface for the user to interact with during operaton There can be several implementations of it dependinng on communication model

Required methods

fn connect(
&self,
target: TransportMethod
) -> Result<Option<PeerId>, SocketError>

Connect to a new peer specified by the parameter. The result can contain the PeerId which should be used in MessageMetadata in future transactions to identify the peer

fn bind(&self, target: TransportMethod) -> Result<Option<PeerId>, SocketError>

Bind to an interface to listen for connections given by the first parameter. The result can contain a PeerId that will contain the self identifier PeerId.

fn close_connection(
&self,
peer_identification: PeerIdentification
) -> Result<(), SocketError>

Close connection to a peer specified by PeerIdentification

fn close(self) -> Result<(), SocketError>

Close the socket with all its connections

Loading content...