Trait rustymq::base::transport::Transport

pub trait Transport: Send + Sync {
    fn send(
        &self, 
        message: RawMessage, 
        flags: OpFlag
    ) -> Result<(), SocketError>;

    fn receive(
        &self, 
        flags: OpFlag
    ) -> Result<RawMessage, (Option<PeerId>, SocketError)>;

    fn close_connection(
        &self, 
        peer_identification: PeerIdentification
    ) -> Result<Option<PeerId>, SocketError>;

    fn query_connected_peers(&self) -> HashSet<PeerId>;

    fn query_configuration(&self) -> Option<&TransportConfiguration>;

    fn close(self) -> Result<(), SocketError>;
}

Transport

Trait that every transport layer implementation has to implement. This is the interface a Socket uses internally to establis a connection

Required methods

fn send(&self, message: RawMessage, flags: OpFlag) -> Result<(), SocketError>

Send message

fn receive(
&self,
flags: OpFlag
) -> Result<RawMessage, (Option<PeerId>, SocketError)>

Receive message

fn close_connection(
&self,
peer_identification: PeerIdentification
) -> Result<Option<PeerId>, SocketError>

Close connection to the peer identified by PeerIdentification

fn query_connected_peers(&self) -> HashSet<PeerId>

Query the PeerIds of all the connected peers

fn query_configuration(&self) -> Option<&TransportConfiguration>

Query the stored configuration if present

fn close(self) -> Result<(), SocketError>

Close the connection to all connected peers

Loading content...