Struct rustymq::core::message::MessageMetadata

pub struct MessageMetadata { /* fields omitted */ }

Metadata for every ZeroMQ Message

This datastructure contains data for tracking, and routing messages between links and the built up communication network

Conained metadata

Randomly generated message indentifier for tracking messages

Randomly generated conversation identifier for tracking conversations

Model ID handled by the communication models used. For defails see super::super::model:here

Peer ID for identifying the peer the message is addressed to or received from. Optionality is dependent on the communication model used. For details see super::super::model:here

Optional multipart message indexing

Implementations

impl MessageMetadata

pub fn new() -> Self

Create a new randomly generated MessageMetadata instance

Message identifier randomly generated

Conversation identifier randomly generated

Model identifier set to None

Peer identifier set to None

Part tracker set to None

Example

let metadata = MessageMetadata::new();

assert!(std::matches!(metadata.peer_id(), None));
assert!(std::matches!(metadata.part(), Part::Single));

pub fn new_multipart() -> Self

Create a new randomly generated MessageMetadata instance for tracking a multipart message

Message identifier randomly generated

Conversation identifier randomly generated

Model identifier set to None

Peer identifier set to None

Part tracker set to Part::Intermediate(0)

Example

let metadata = MessageMetadata::new_multipart();

assert!(std::matches!(metadata.peer_id(), None));
assert!(std::matches!(metadata.part(), Part::Intermediate(0)));

pub fn applied_peer_id(self, id: PeerId) -> Self

Generate new instance from current one with a peer identifier applied

This method will commit the peer identifier given to the metadata but retain all the other fields

Example

let metadata = MessageMetadata::new();
assert!(std::matches!(metadata.peer_id(), None));

let peer_id = PeerId::new(5);
let metadata = metadata.applied_peer_id(peer_id.clone());
assert!(std::matches!(metadata.peer_id(), Some(peer_id)));

pub fn started_multipart(self) -> Self

Generate new instance from current one with multipart message part tracker

This method will set the part tracking of the metadata to Some(Part::Intermediate(0)) while retaining all other fields

Example

let metadata = MessageMetadata::new();
assert!(std::matches!(metadata.part(), Part::Single));

let metadata = metadata.started_multipart();
assert!(std::matches!(metadata.part(), Part::Intermediate(0)));

pub fn continue_exchange(self) -> Self

Generate new instance for continuing an already existing conversation

This method will retain all fields of the metadata except for the message identifier and part

Example

let original = MessageMetadata::new();
let original_converstion_id = original.conversation_id().clone();
let original_clone = original.clone();

let continuation = original.continue_exchange();
assert_ne!(original_clone.message_id(), continuation.message_id());
assert_eq!(original_clone.conversation_id(), continuation.conversation_id());
assert_eq!(original_clone.peer_id(), continuation.peer_id());

pub fn commit_communication_model_id(
self,
model_id: CommunicationModelId
) -> Self

Generate new instace with a new communication model id. (Only needed for custom communication model implementation)

This is only needed in case you are implementing your own communication model. It will retain all fields of the metadata except for the communication model identifier given as the parameter

pub fn message_id(&self) -> &MessageId

Qurey the stored message id

pub fn conversation_id(&self) -> &ConversationId

Query the stored convesation id

pub fn communication_model_id(&self) -> &Option<CommunicationModelId>

Query the stored communication model id (Only needed for custom communication model implementation)

pub fn peer_id(&self) -> &Option<PeerId>

Query the stored peer identifier (None if not set)

pub fn is_multipart(&self) -> bool

Query whether the message is multipart

pub fn part(&self) -> &Part

Query the part tracker (None if not set)

pub fn next_multipart(self) -> Result<Self, PartError>

Generate new instance for multipart message metadata with advancing the part id by one

Example

let original = MessageMetadata::new_multipart();
assert!(std::matches!(original.part(), Part::Intermediate(0)));

let result = original.next_multipart().unwrap();
assert!(std::matches!(result.part(), Part::Intermediate(1)));

pub fn as_final_multipart(self) -> Result<Self, PartError>

Generate new instance for multipart message metadata without advancing the part id and changing it to final

Example

let original = MessageMetadata::new_multipart();
assert!(std::matches!(original.part(), Part::Intermediate(0)));

let result = original.as_final_multipart().unwrap();
assert!(std::matches!(result.part(), Part::Final(0)));

pub fn next_final_multipart(self) -> Result<Self, PartError>

Generate new instance for multipart message metadata with advancing the part id and changing it to final

Example

let original = MessageMetadata::new_multipart();
assert!(std::matches!(original.part(), Part::Intermediate(0)));

let result = original.as_final_multipart().unwrap();
assert!(std::matches!(result.part(), Part::Final(0)));