Trait rustymq::core::serializer::Serializable

pub trait Serializable where
    Self: Sized, {
    fn serialize<T: Serializer>(&self, serializer: &mut T);

    fn deserialize<T: Deserializer>(deserializer: &mut T) -> Result<Self, Error>;
}

Trait that enables a Serializer object can serialize and a Deserializer object can deserialize

Required methods

fn serialize<T: Serializer>(&self, serializer: &mut T)

Serialize self into serializer

fn deserialize<T: Deserializer>(deserializer: &mut T) -> Result<Self, Error>

Deserialize a value with the type of Self from a deserializer

Loading content...