Trait rustymq::core::socket::InwardSocket

pub trait InwardSocket: Socket {
    fn receive(
        &self, 
        flags: OpFlag
    ) -> Result<RawMessage, (Option<PeerId>, SocketError)>;

    fn receive_typed<T>(
        &self, 
        flags: OpFlag
    ) -> Result<TypedMessage<T>, ReceiveTypedError<T>>
    where
        T: SerializableMessagePayload,
    { ... }
}

Inward Socket trait

The main interface for the user to receive messages through the opened RustyMQ socket There can be several implementations of it depending on the communication model used

Required methods

fn receive(
&self,
flags: OpFlag
) -> Result<RawMessage, (Option<PeerId>, SocketError)>

Receive a raw message

Loading content...
Provided methods

fn receive_typed<T>(
&self,
flags: OpFlag
) -> Result<TypedMessage<T>, ReceiveTypedError<T>> where
T: SerializableMessagePayload,

Receive a typed message. Requires T to implement the SerializableMessagePayload trait

Loading content...