Module rustymq::core::util::thread

Threaded operation helper functionality

Timing functionality used internally by RustyMQ

Structs

ChangeNotifyMutex

ChangeNotifyMutexGuard

Semaphore
Semaphore

Sleeper
#Sleeper A helper struct that allows for sleeping the currrent thread according to a duration backoff algorithm

Functions

poll
Execute operation repeadetly until success with sleep_time sleep between attemps

poll_backoff
Execute operation repeadetly until success with a Sleeper sleep between each attempt

poll_backoff_mut
Execute operation repeadetly until success with a Sleeper sleep between each attempt

poll_mut
Execute operation repeadetly until success with sleep_time sleep between attemps

sleep

Type Definitions

ChgNtfMutex

ChgNtfMutexGuard