Module rustymq::internals::serializer

Serializer module

This module contains a basic serialization functionality used internally by RusyMQ But it can be used externally for serialization as a basic tool for serialization.

Details

The reason not to use a serialization library was to keep the dependencies of the library minimal and also to have a simple and fast binary serialization that also supports big and little endiannes

Structs

FlatDeserializer
Implementation for Deserializer that restores objects from a flat binary buffer

FlatSerializer
Implementation for Serializer that stores values in a flat binary buffer

Enums

Error
Signaling serialization errors

Traits

Deserializer
Deserializer trait that can deserialize a buffer containining Serializable objects

RawSerializable
Trait that marks primitive types that can be serialized with Serializer object

Serializable
Trait that enables a Serializer object can serialize and a Deserializer object can deserialize

Serializer
Serializer trait that can serialize a Serializable object

Type Definitions

Buffer
General buffer type to be used in RusyMQ

BufferMutSlice
General buffer mutable slice type to be used in RustyMQ

BufferSlice
General buffer slice type to be used in RustyMQ