Trait rustymq::internals::serializer::Deserializer

pub trait Deserializer: Sized {
    fn consume<'a>(
        &'a mut self, 
        amount: usize
    ) -> Result<BufferSlice<'a>, Error>;

    fn deserialize<T: Serializable>(&mut self) -> Result<T, Error>;

    fn byte_order_correction(&self) -> bool;

    fn deserialize_raw<T: RawSerializable>(&mut self) -> Result<T, Error> { ... }

    fn deserialize_raw_slice<T: RawSerializable>(
        &mut self
    ) -> Result<Vec<T>, Error> { ... }
}

Deserializer trait that can deserialize a buffer containining Serializable objects

Required methods

fn consume<'a>(&'a mut self, amount: usize) -> Result<BufferSlice<'a>, Error>

Consume an amount of the the buffer being parsed by the deserializer

fn deserialize<T: Serializable>(&mut self) -> Result<T, Error>

Deserialize a Serializable object from the buffer being parsed by the deserializer

fn byte_order_correction(&self) -> bool

Check if byte order correction is needed on RawSerializable values being deserialized

Loading content...
Provided methods

fn deserialize_raw<T: RawSerializable>(&mut self) -> Result<T, Error>

Deserialize a RawSerializable value from the buffer being parsed by the deserializer

fn deserialize_raw_slice<T: RawSerializable>(&mut self) -> Result<Vec<T>, Error>

Deserialize a slice of RawSerializable values from the buffer being parsed by the deserializer

Loading content...