Trait rustymq::internals::serializer::Serializer

pub trait Serializer: Sized {
    fn append<'a>(&mut self, slice: BufferSlice<'a>);

    fn finalize(self) -> Buffer;

    fn serialize<T: Serializable>(&mut self, serializable: &T) { ... }

    fn serialize_pass<T: Serializable>(&mut self, serializable: T) { ... }

    fn serialize_raw<T: RawSerializable>(&mut self, serializable: &T) { ... }

    fn serialize_raw_slice<T: RawSerializable>(&mut self, slice: &[T]) { ... }
}

Serializer trait that can serialize a Serializable object

Required methods

fn append<'a>(&mut self, slice: BufferSlice<'a>)

Append a slice slice to the buffer stored in the serializer

fn finalize(self) -> Buffer

Finalize the serializer and return the buffer containing the serialized data

Loading content...
Provided methods

fn serialize<T: Serializable>(&mut self, serializable: &T)

Serialize Serializable object by reference

fn serialize_pass<T: Serializable>(&mut self, serializable: T)

Serialize Serializable object by passing ownership

fn serialize_raw<T: RawSerializable>(&mut self, serializable: &T)

Serizalize RawSerializable value

fn serialize_raw_slice<T: RawSerializable>(&mut self, slice: &[T])

Serizalize a slice of RawSerializable objects

Loading content...