Struct rustymq::model::pubsub::PublisherSocket

pub struct PublisherSocket<T: AcceptorTransport> { /* fields omitted */ }

Publisher socket

This socket can be used to distribute information to subscribing peers

Implementations

impl<T> PublisherSocket<T> where
T: AcceptorTransport,

pub fn new(transport: T) -> Self

Creates new socket with the underlieing transport to distribute messages to subscribint peers