Module rustymq::model::reqrep

Request reply communication model

This module contains the socket implementation for a request-reply communication model.

Operation

The ReplySocket listens for incoming connections and shuld reply to all incoming messages. The RequestSocket can be used to send requests to the reply sockets.

Example

use rustymq::base::{Socket, InwardSocket, OutwardSocket, OpFlag, BidirectionalSocket,
                    Message, RawMessage,
                    TransportMethod, NetworkAddress};
use rustymq::model::{RequestSocket, ReplySocket};
use rustymq::transport::network::tcp;

let replier = ReplySocket::new(
                tcp::AcceptorTransport::new(
                    tcp::StreamConnectionBuilder::new(),
                    tcp::StreamListenerBuilder::new()))
              .unwrap();
let requestor = RequestSocket::new(tcp::InitiatorTransport::new(tcp::StreamConnectionBuilder::new())).unwrap();

replier.bind(NetworkAddress::from_dns("localhost:13000".to_string()).unwrap().into());
requestor.connect(NetworkAddress::from_dns("localhost:13000".to_string()).unwrap().into());

// Make sure that TCP connection is established otherwise the first sent message might not arrive
std::thread::sleep(std::time::Duration::from_millis(100));

let request_payload: Vec<u8> = vec![2u8, 8u8];
let response_payload_requirement: Vec<u8> = vec![2u8, 8u8, 16u8, 32u8];

requestor.send(RawMessage::new(request_payload.clone()), OpFlag::Wait);
replier.respond(OpFlag::Wait, OpFlag::Wait, |message| {
    let mut payload = message.into_payload();
    let reply_payload_extend: Vec<u8> = vec![16u8, 32u8];
    payload.extend(reply_payload_extend.into_iter());
    RawMessage::new(payload)
});
let response_payload = requestor.receive(OpFlag::Wait).unwrap().into_payload();

assert_eq!(response_payload.clone(), response_payload_requirement.clone());

Structs

ReplySocket
Reply socket

RequestSocket
Request socket